+++ src/main/java/edu/hm/webtech/topic/TopicRepository.java
@@ -1,10 +1,7 @@
-package edu.hm.webtech.repositories;
+package edu.hm.webtech.topic;
 
-import edu.hm.webtech.entities.Topic;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.List;
 
 @Repository
 public interface TopicRepository extends CrudRepository<Topic, Long> {
