+++ src/test/java/edu/hm/webtech/ClozeTest.java
@@ -3,11 +3,13 @@ package edu.hm.webtech;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import edu.hm.webtech.entities.Cloze;
-import edu.hm.webtech.entities.Topic;
-import edu.hm.webtech.repositories.ClozeRepository;
-import edu.hm.webtech.repositories.ExerciseRepository;
-import edu.hm.webtech.repositories.TopicRepository;
+import edu.hm.webtech.cloze.Cloze;
+import edu.hm.webtech.exercise.BloomLevel;
+import edu.hm.webtech.exercise.TopicBloomLevel;
+import edu.hm.webtech.topic.Topic;
+import edu.hm.webtech.cloze.ClozeRepository;
+import edu.hm.webtech.exercise.ExerciseRepository;
+import edu.hm.webtech.topic.TopicRepository;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
