+++ src/test/java/edu/hm/webtech/TopicTest.java
@@ -3,8 +3,8 @@ package edu.hm.webtech;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.ObjectWriter;
 import com.fasterxml.jackson.databind.SerializationFeature;
-import edu.hm.webtech.entities.Topic;
-import edu.hm.webtech.repositories.TopicRepository;
+import edu.hm.webtech.topic.Topic;
+import edu.hm.webtech.topic.TopicRepository;
 import org.hamcrest.core.IsNull;
 import org.junit.After;
 import org.junit.Before;
